Drop support for iptables 1.8.3, update distro version notes
iptables 1.8.3 has a bug that makes it hang forever in certain cases,
requiring us to use "timeout" in the script. However, 1.8.4 is over
two years old at this point, and even Debian buster-backports has
1.8.5, so there's no reason to keep supporting 1.8.3. So remove that.

Also update the README and test Dockerfiles to use newer versions of
Fedora and Alpine to get a new-enough version.