Convert iptables-wrapper to a static go program
Using a static binary removes the need to include sh in the kube-proxy
image. It also drops the dependency on grep and wc.

The binary is about 1.7MB in linux amd64. This will make the resulting
image bigger, although removing dash, grep and wc will compensate for
that.

The Go code is almost a drop-in replacement for the existing sh script.
The only differences are:
* The Go program uses a new process to execute the proxied iptables
  command instead of replacing the current process with exec. It
  redirects stderr and stdout so this should be equivalent.
* iptables binary directory is detected on runtime. The cost should me
  minimum, it just looks for a file in two folders.
* The system iptables selector (alternatives vs update-alternatives) is
  detected in runtime as opposed to during wrapper installation. The
  cost should be minimum, it just inspects the sbin directory for one
  binary or the other.

Co-authored-by: Dan Winship <danwinship@redhat.com>