Ignore redirection error if target already exists
Do not log an error if redirection failed because the target already
exists. This can happen when multiple iptables-wrappers processes race
each other, or the filesystem is mounted read-only (and removing the
existing symlink silently failed). There is nothing to be done in these
cases and the logs are just noise.